requirements.txt
pytorch
sentence-transformers
pandas
scipy
openai==0.28.0
openpyxl
scikit-learn
tabulate
defopt
pyyaml
ray